Refactor Livewire components with computed properties and caching.
Replaced property access with computed attributes to optimize queries and enable caching. Added cache invalidation for `published-count` upon article creation, update, or deletion. Streamlined query handling for `articles` and updated corresponding Blade templates.

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -14,9 +15,29 @@ class ArticleList extends AdminComponent
 
     public $showOnlyPublished = false;
 
+//    Using the Computed attribute creates a computed property to allow a query of articles to be cached for use within a livewire request.
+    #[Computed]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', true);
+        }
+
+        return $query->paginate(10, pageName: 'articles-page');
+    }
+
     public function delete(Article $article)
     {
+        if ($this->articles->count() < 10) {
+            throw new \Exception('Nope');
+        }
+
         $article->delete();
+
+        unset($this->articles);
+        cache()->forget('published-count');
     }
 
     public function showAll()
@@ -33,14 +54,6 @@ class ArticleList extends AdminComponent
 
     public function render()
     {
-        $query = Article::query();
-
-        if ($this->showOnlyPublished) {
-            $query->where('published', true);
-        }
-
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'articles-page'),
-        ]);
+        return view('livewire.article-list');
     }
 }
